Added a lot. Made the add future event button work, made some basic checkers to confirm the fields are all in correct. Made the save button work. Now outputs all the data for that day and only prints them once. (need to connect to database so, it will save for the future). Changed the labels, so they all fade out like the save label does. Changed some FXML files a bit. Removed some redundant code from my files.
Current known issue: No technically an issue it works as intended, but it's a problem. If an event goes for more than 1 day then the save button will register that event for each of the days it spans on. At 3am I couldn't figure out a solution so hopefully someone can if not it will have to be done on the database side making the entry unique.